Updated the errors

+++ contracts/src/MorphValidatorVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract MorphValidatorVerifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 3050421639393894609694100247362469476301594184194208660243717647327937608084;
-    uint256 constant alphay  = 10570968496776547372902278874938397270085925787545262165721065498216990788862;
-    uint256 constant betax1  = 21691111980123012323298500290421361848112563758657263962991588788225312145784;
-    uint256 constant betax2  = 4216856901912371297629069257509223183173062360485498483168636816601708818350;
-    uint256 constant betay1  = 4563169662782910687268173125798006935932702707712370780282685207883966408839;
-    uint256 constant betay2  = 20270793118337942609509108123032858522765806053947340581216635283059651491912;
+    uint256 constant alphax = 3050421639393894609694100247362469476301594184194208660243717647327937608084;
+    uint256 constant alphay = 10570968496776547372902278874938397270085925787545262165721065498216990788862;
+    uint256 constant betax1 = 21691111980123012323298500290421361848112563758657263962991588788225312145784;
+    uint256 constant betax2 = 4216856901912371297629069257509223183173062360485498483168636816601708818350;
+    uint256 constant betay1 = 4563169662782910687268173125798006935932702707712370780282685207883966408839;
+    uint256 constant betay2 = 20270793118337942609509108123032858522765806053947340581216635283059651491912;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,30 +42,33 @@ contract MorphValidatorVerifier {
     uint256 constant deltay1 = 4765254222855890982565417673590547699100913701246833672730702205957238474279;
     uint256 constant deltay2 = 14983805389688528635752010507989002925519714062335005215662480695023594693286;
 
-    
     uint256 constant IC0x = 7410883609919575672632697569957737350095025710095515214385744816797544934154;
     uint256 constant IC0y = 5050477155140491416373758599654288067509939069487301420790413777560120293113;
-    
+
     uint256 constant IC1x = 9197873625030438261372385069915494882493763348487114722197796082403675139162;
     uint256 constant IC1y = 15424144416123496373956751658625636489752927243355267691242792971384574273100;
-    
+
     uint256 constant IC2x = 19370023096041004709084254280919499284013037871939430548190029800008360400073;
     uint256 constant IC2y = 3178407707658334056363798135371619262218393677764055495147332642960362648646;
-    
+
     uint256 constant IC3x = 19747224303349628823402836119037158916698346392194956518523260061112369348192;
     uint256 constant IC3y = 21356810064174330540658007319508165803399438376990332592285198653401705562567;
-    
+
     uint256 constant IC4x = 2973025877662129610718903563882235995198967245729189078791484937373487162130;
     uint256 constant IC4y = 16033438617694528902829542235048953834119488851860687539470253437144780864641;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[4] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[4] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -73,7 +76,7 @@ contract MorphValidatorVerifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -108,15 +111,14 @@ contract MorphValidatorVerifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_pVk, IC4x, IC4y, calldataload(add(pubSignals, 96)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -142,7 +144,6 @@ contract MorphValidatorVerifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -159,7 +160,6 @@ contract MorphValidatorVerifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -169,21 +169,20 @@ contract MorphValidatorVerifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
