+++ contracts/src/MorphVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity 0.8.20;
 
 contract MorphVerifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant ALPHAX  = 3050421639393894609694100247362469476301594184194208660243717647327937608084;
-    uint256 constant ALPHAY  = 10570968496776547372902278874938397270085925787545262165721065498216990788862;
-    uint256 constant BETAX1  = 21691111980123012323298500290421361848112563758657263962991588788225312145784;
-    uint256 constant BETAX2  = 4216856901912371297629069257509223183173062360485498483168636816601708818350;
-    uint256 constant BETAY1  = 4563169662782910687268173125798006935932702707712370780282685207883966408839;
-    uint256 constant BETAY2  = 20270793118337942609509108123032858522765806053947340581216635283059651491912;
+    uint256 constant ALPHAX = 3050421639393894609694100247362469476301594184194208660243717647327937608084;
+    uint256 constant ALPHAY = 10570968496776547372902278874938397270085925787545262165721065498216990788862;
+    uint256 constant BETAX1 = 21691111980123012323298500290421361848112563758657263962991588788225312145784;
+    uint256 constant BETAX2 = 4216856901912371297629069257509223183173062360485498483168636816601708818350;
+    uint256 constant BETAY1 = 4563169662782910687268173125798006935932702707712370780282685207883966408839;
+    uint256 constant BETAY2 = 20270793118337942609509108123032858522765806053947340581216635283059651491912;
     uint256 constant GAMMAX1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant GAMMAX2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant GAMMAY1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,30 +42,33 @@ contract MorphVerifier {
     uint256 constant DELTAY1 = 17940180404375462993749773111150630102443424426314010844370940044289770093360;
     uint256 constant DELTAY2 = 21221829802419819553697275816550557864537352963471280606040202401359281973141;
 
-    
     uint256 constant IC0X = 4330896101261327931562894222304268907005630367277661418240976629349999434;
     uint256 constant IC0Y = 19997158488710195389006109539486318001138730795830822891269578160878833292296;
-    
+
     uint256 constant IC1X = 5884790233945144456835786709185480620191886079201270112598989400779947080097;
     uint256 constant IC1Y = 7753638646293425949442237084470669943195647866179890881666447434852366108174;
-    
+
     uint256 constant IC2X = 3111468173117182385088039053307551782915738539700109360203239853912727799418;
     uint256 constant IC2Y = 1215203918928959422692924005958635099664657458128909371858839425066204742317;
-    
+
     uint256 constant IC3X = 2297451708450523864156881426943996729308355969589882314526557016000994226674;
     uint256 constant IC3Y = 3678370412405661836157728963973702620532096259993717115138320768283660304490;
-    
+
     uint256 constant IC4X = 387256210787584577913315239850335881940610232554516473662530369917182383073;
     uint256 constant IC4Y = 5834205663288724696918314961286589077794243120413574472177033587451436867947;
-    
- 
+
     // Memory data
     uint16 constant P_VK = 0;
     uint16 constant PPAIRING = 128;
 
     uint16 constant PLASTMEM = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[4] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[4] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -73,7 +76,7 @@ contract MorphVerifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -108,15 +111,14 @@ contract MorphVerifier {
                 mstore(add(_P_VK, 32), IC0Y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_P_VK, IC1X, IC1Y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_P_VK, IC2X, IC2Y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_P_VK, IC3X, IC3Y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_P_VK, IC4X, IC4Y, calldataload(add(pubSignals, 96)))
-                
 
                 // -A
                 mstore(_PPAIRING, calldataload(pA))
@@ -142,7 +144,6 @@ contract MorphVerifier {
                 mstore(add(_PPAIRING, 384), mload(add(pMem, P_VK)))
                 mstore(add(_PPAIRING, 416), mload(add(pMem, add(P_VK, 32))))
 
-
                 // gamma2
                 mstore(add(_PPAIRING, 448), GAMMAX1)
                 mstore(add(_PPAIRING, 480), GAMMAX2)
@@ -159,7 +160,6 @@ contract MorphVerifier {
                 mstore(add(_PPAIRING, 704), DELTAY1)
                 mstore(add(_PPAIRING, 736), DELTAY2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _PPAIRING, 768, _PPAIRING, 0x20)
 
                 isOk := and(success, mload(_PPAIRING))
@@ -169,21 +169,20 @@ contract MorphVerifier {
             mstore(0x40, add(pMem, PLASTMEM))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
