+++ contracts/src/StateCommitmentGroth16Verifier.sol
@@ -22,17 +22,17 @@ pragma solidity ^0.8.20;
 
 contract StateCommitmentGroth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 3050421639393894609694100247362469476301594184194208660243717647327937608084;
-    uint256 constant alphay  = 10570968496776547372902278874938397270085925787545262165721065498216990788862;
-    uint256 constant betax1  = 21691111980123012323298500290421361848112563758657263962991588788225312145784;
-    uint256 constant betax2  = 4216856901912371297629069257509223183173062360485498483168636816601708818350;
-    uint256 constant betay1  = 4563169662782910687268173125798006935932702707712370780282685207883966408839;
-    uint256 constant betay2  = 20270793118337942609509108123032858522765806053947340581216635283059651491912;
+    uint256 constant alphax = 3050421639393894609694100247362469476301594184194208660243717647327937608084;
+    uint256 constant alphay = 10570968496776547372902278874938397270085925787545262165721065498216990788862;
+    uint256 constant betax1 = 21691111980123012323298500290421361848112563758657263962991588788225312145784;
+    uint256 constant betax2 = 4216856901912371297629069257509223183173062360485498483168636816601708818350;
+    uint256 constant betay1 = 4563169662782910687268173125798006935932702707712370780282685207883966408839;
+    uint256 constant betay2 = 20270793118337942609509108123032858522765806053947340581216635283059651491912;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,24 +42,27 @@ contract StateCommitmentGroth16Verifier {
     uint256 constant deltay1 = 19909797708976680321698327758256773146938768861163668319262642820375488993402;
     uint256 constant deltay2 = 7684090622620134716676444437712651127319803356773660826784650502880159849139;
 
-    
     uint256 constant IC0x = 8897313170330274008474300941622954799324914180233253735564879494219603825055;
     uint256 constant IC0y = 18865597183192934992363832004951145266986258405709180176023449174446228440145;
-    
+
     uint256 constant IC1x = 17911283209436326003686978313783819385669339395707766403480353807511159948836;
     uint256 constant IC1y = 10660392718687467573128694424483648029379505495213688982398460117716668183443;
-    
+
     uint256 constant IC2x = 8052522557305527945084726096364887111243009224027803751438954856440159310269;
     uint256 constant IC2y = 64305206098522302854330047280891354758008347834376308701362775698938106094;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[2] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[2] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -67,7 +70,7 @@ contract StateCommitmentGroth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -102,11 +105,10 @@ contract StateCommitmentGroth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -132,7 +134,6 @@ contract StateCommitmentGroth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -149,7 +150,6 @@ contract StateCommitmentGroth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -159,17 +159,16 @@ contract StateCommitmentGroth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
